# SILO-plateforme-source-2026-07-27/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 3C Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      description: Admins see all projects; clients see their own
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [deliverables]
      summary: List private project resources visible to a project member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: addProjectFile
      tags: [deliverables]
      summary: Register an uploaded private project resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileInput"
      responses:
        "201":
          description: Project file registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "403":
          description: Forbidden
...
  /partners:
    get:
      operationId: listPartners
      tags: [partner]
      summary: List partners in the caller's scope
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partner]
      summary: Create a partner profile (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Partner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    patch:
      operationId: updatePartner
      tags: [partner]
      summary: Update a partner profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Partner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
...
  /comm/projects/{projectId}:
    get:
      operationId: getProjectCommStatus
      tags: [comm]
      summary: Communication status for a project (shaped by role)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project communication status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommProjectStatus"
...
components:
  schemas:
...
    CommProjectStatus:
      type: object
      required: [projectId, directEnabled]
      properties:
        projectId:
          type: string
        adminAutorisePm:
          type: ["boolean", "null"]
        activeParPm:
          type: ["boolean", "null"]
        blocageGlobal:
          type: ["boolean", "null"]
        directEnabled:
          type: boolean
...
    Project:
      type: object
      required: [id, title, type, status, clientName, clientEmail, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [mariage, clip, corporate, reseaux]
        status:
          type: string
          enum:
            [
              lead,
              devis,
              production,
              post_production,
              livraison,
              termine,
              annule,
            ]
        clientName:
          type: string
        clientEmail:
          type: string
        clientUserId:
          type: ["string", "null"]
        advisorUserId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        shootingDate:
          type: ["string", "null"]
          format: date
        deliveryDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title, type, clientName, clientEmail]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [mariage, clip, corporate, reseaux]
        clientName:
          type: string
        clientEmail:
          type: string
        clientUserId:
          type: ["string", "null"]
        advisorUserId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        shootingDate:
          type: ["string", "null"]
          format: date
        deliveryDate:
          type: ["string", "null"]
          format: date
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [mariage, clip, corporate, reseaux]
        status:
          type: string
          enum:
            [
              lead,
              devis,
              production,
              post_production,
              livraison,
              termine,
              annule,
            ]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        shootingDate:
          type: ["string", "null"]
          format: date
        deliveryDate:
          type: ["string", "null"]
          format: date
        clientUserId:
          type: ["string", "null"]
        advisorUserId:
          type: ["string", "null"]
...
    ProjectFile:
      type: object
      required:
        [
          id,
          projectId,
          name,
          storageBucket,
          storagePath,
          sizeBytes,
          sourceRole,
          createdAt,
        ]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        storageBucket:
          type: string
          enum: [project-files, client-uploads]
        storagePath:
          type: string
        mimeType:
          type: ["string", "null"]
        sizeBytes:
          type: integer
          minimum: 0
        sourceRole:
          type: string
          enum: [client, pm, admin]
        createdAt:
          type: string
          format: date-time
    ProjectFileInput:
      type: object
      required: [name, storageBucket, storagePath, sizeBytes]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 500
        storageBucket:
          type: string
          enum: [project-files, client-uploads]
        storagePath:
          type: string
          minLength: 1
          maxLength: 2000
        mimeType:
          type: ["string", "null"]
          maxLength: 255
        sizeBytes:
          type: integer
          minimum: 0
...
    Partner:
      type: object
      required:
        [
          id,
          userId,
          name,
          kind,
          email,
          specialties,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [agency, individual]
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        portfolioUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, suspended, rejected]
        advisorUserId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PartnerInput:
      type: object
      required: [userId, name, kind, email]
      properties:
        userId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [agency, individual]
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        portfolioUrl:
          type: ["string", "null"]
        advisorUserId:
          type: ["string", "null"]
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        portfolioUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, suspended, rejected]
        advisorUserId:
          type: ["string", "null"]
...